finishing practice fusion scraper

+++ practice_fusion_scraper.py
@@ -1,8 +1,11 @@
 from selenium import webdriver
-from selenium.webdriver.common.action_chains import ActionChains
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
+
+from pdf2image import convert_from_path
+from PIL import Image
+import pytesseract
 
 import difflib
 import datetime
@@ -13,6 +16,7 @@ import os
 from time import sleep
 
 from config import config
+pytesseract.pytesseract.tesseract_cmd = config['tesseract_path']
 
 
 def practice_fusion_login(driver: webdriver.Chrome):
@@ -34,6 +38,7 @@ def practice_fusion_login(driver: webdriver.Chrome):
     return two_factor_authentication(driver)
 
 
+# TODO: if 2fa code is wrong
 def two_factor_authentication(driver: webdriver.Chrome) -> bool:
     """Helps automate the 2fa process
 
@@ -84,7 +89,6 @@ def get_patient_full_name(patient_data, row: int) -> str:
     return patient_full_name
 
 
-# TODO: error checking/report on patient names
 def navigate_to_patient_info(driver: webdriver.Chrome, full_name: str) -> bool:
     search_patient_first_names: list[WebElement] = driver.find_elements(
         By.CSS_SELECTOR, config['pf_first_name_css_select'])
@@ -97,10 +101,14 @@ def navigate_to_patient_info(driver: webdriver.Chrome, full_name: str) -> bool:
         search_full_name: str = f'{search_first_name} {search_last_name}'
         full_name_element_mapping[search_full_name] = first_name_ele
 
-    closest_match_name: str = difflib.get_close_matches(
+    closest_matches: str = difflib.get_close_matches(
         word=full_name,
-        possibilities=full_name_element_mapping.keys())[0]
-    closest_match_element: WebElement = full_name_element_mapping[closest_match_name]
+        possibilities=full_name_element_mapping.keys(),
+        cutoff=0.75)
+    if len(closest_matches) == 0:
+        return False
+
+    closest_match_element: WebElement = full_name_element_mapping[closest_matches[0]]
     closest_match_element.click()
     sleep(1) # wait for the webpage to load
 
@@ -126,13 +134,19 @@ def navigate_to_profile_tab(driver: webdriver.Chrome):
     sleep(1)
 
 
-def get_patient_address(driver: webdriver.Chrome):
-    address_line_1: str = driver.find_element(By.CSS_SELECTOR, config['pf_address1_css_select']).text
-    address_line_2: str = driver.find_element(By.CSS_SELECTOR, config['pf_address2_css_select']).text
-    city_state_zip: str = driver.find_element(By.CSS_SELECTOR, config['pf_address3_css_select']).text
-    
+def get_patient_address(driver: webdriver.Chrome) -> tuple[str, str, str, str] | bool:
+    try:
+        address_line_1: str = driver.find_element(By.CSS_SELECTOR, config['pf_address1_css_select']).text
+        address_line_2: str = driver.find_element(By.CSS_SELECTOR, config['pf_address2_css_select']).text
+        city_state_zip: str = driver.find_element(By.CSS_SELECTOR, config['pf_address3_css_select']).text
+    except:
+        return False
+
     city_state_zip = city_state_zip.replace(',', '')
-    city, state, zip_code = city_state_zip.split(' ')
+    address_3_parts = city_state_zip.split(' ')
+    state = address_3_parts[-2]
+    zip_code = address_3_parts[-1]
+    city = ' '.join(address_3_parts[0:-2])
 
     return address_line_1, address_line_2, city, state, zip_code
 
@@ -144,31 +158,40 @@ def get_patient_sex(driver: webdriver.Chrome):
     return sex
 
 
-def navigate_to_desired_document(driver: webdriver.Chrome, operation: str, operation_date: datetime.datetime):
-    documents: list[WebElement] = driver.find_elements(By.CSS_SELECTOR, config['pf_document_info_css_select'])
+def navigate_to_desired_document(driver: webdriver.Chrome, operation: str, operation_date: datetime.datetime) -> bool:
+    documents_container: WebElement = driver.find_element(By.CSS_SELECTOR, config['pf_document_container_css_select'])
+    documents: list[WebElement] = documents_container.find_elements(By.XPATH, config['pf_document_row_xpath'])
+    document_found: bool = False
 
     for document in documents:
-        document_type: WebElement = document.find_element(By.CSS_SELECTOR, config['pf_document_type_css_select'])
-        document_type_str: str = document_type.find_element(By.XPATH, config['pf_document_type_xpath']).text
+        document_information_containers: list[WebElement] = document.find_elements(By.XPATH, config['xpath_child_selector'])
         
+        document_type: WebElement = document_information_containers[2]
+        document_type_div: WebElement = document_type.find_element(By.XPATH, config['xpath_child_selector'])
+        document_type_str: str = document_type_div.find_element(By.XPATH, config['xpath_child_selector']).text
+
         if operation.strip().lower() not in document_type_str.lower():
             continue
 
-        document_date_str: str = document.find_element(By.XPATH, config['pf_document_date_xpath']).text
+        document_date_str: WebElement = document.find_elements(By.XPATH, config['xpath_child_selector'])[5].text
         document_date: datetime.datetime = datetime.datetime.strptime(document_date_str, '%m/%d/%Y')
-        end_date_range: datetime.datetime = operation_date + datetime.timedelta(days=5)
+        end_date_range: datetime.datetime = operation_date + datetime.timedelta(days=10)
 
         if not (operation_date <= document_date <= end_date_range):
             continue
 
-        document_link_element: WebElement = document.find_element(By.XPATH, config['pf_document_link_xpath'])
+        document_link_container: WebElement = document_information_containers[1]
+        document_link_div: WebElement = document_link_container.find_element(By.XPATH, config['xpath_child_selector'])
+        document_link_element: WebElement = document_link_div.find_element(By.XPATH, config['xpath_child_selector'])
         document_link_element.click()
+        document_found = True
         break
 
     sleep(3)
+    return document_found
 
 
-def download_document(driver: webdriver.Chrome, patient_name: str, date: datetime.datetime, operation_type: str):
+def download_document(driver: webdriver.Chrome, patient_name: str, date: datetime.datetime, operation_type: str) -> str:
     print_buttons_container: WebElement = driver.find_element(By.CSS_SELECTOR, config['pf_print_buttons_css_select'])
     download_button: WebElement = print_buttons_container.find_element(By.XPATH, config['pf_download_button_xpath'])
     download_button.click()
@@ -179,16 +202,67 @@ def download_document(driver: webdriver.Chrome, patient_name: str, date: datetim
 
     operation_type = operation_type.strip().lower()
     date_str: str = date.strftime('%Y-%m-%d')
-    new_file_name: str = f'{patient_name}_{date_str}_{operation_type}.pdf'
+    new_file_name: str = f'{date_str}_{patient_name}_{operation_type}.pdf'
 
     old_file_path: str = f'{config["pf_pdf_download_location"]}{pdf_name}'
     new_file_path: str = f'{config["pf_pdf_download_location"]}{new_file_name}'
 
     sleep(3)
 
-    os.rename(old_file_path, new_file_path)
+    if not os.path.exists(new_file_path):
+        os.rename(old_file_path, new_file_path)
+
+    return new_file_path
 
 
+def get_text_from_pdf(pdf_file_path: str) -> str:
+    pdf_pages: list[Image.Image] = convert_from_path(pdf_file_path, 300)
+
+    first_page: Image.Image = pdf_pages[0]
+    first_page_text: str = pytesseract.image_to_string(first_page, lang='eng')
+
+    return first_page_text
+
+
+def parse_pdf_text(pdf_text: str) -> tuple[str, str, str]:
+    pdf_text_by_line: list[str] = pdf_text.split('\n')
+
+    id: str = ''
+    procedure_code: str = ''
+    diagnosis_code: str = ''
+    for line in pdf_text_by_line:
+        if line.startswith('ID:'):
+            id_and_procedure: str = line.split(':')[1].strip()
+            id = id_and_procedure.split(' ')[0]
+        elif line.startswith('Procedure Code:'):
+            proc_code_and_name: str = line.split(':')[1].strip()
+            procedure_code = proc_code_and_name.split(' ')[0]
+        elif line.startswith('Diagnosis code:'):
+            diag_code_and_type: str = line.split(':')[1].strip()
+            diagnosis_code = diag_code_and_type.split(' ')[0]
+
+    return id, procedure_code, diagnosis_code
+
+
+def close_patient_charts_tab(driver: webdriver.Chrome):
+    try:
+        close_charts_container: WebElement = driver.find_element(By.CLASS_NAME, config['pf_close_charts_container_class'])
+        close_charts_button: WebElement = close_charts_container.find_element(By.XPATH, config['pf_close_charts_button_xpath'])
+        close_charts_button.click()
+
+        sleep(3)
+    except:
+        pass
+
+    try:
+        close_dob_filter_button: WebElement = driver.find_element(By.CSS_SELECTOR, config['pf_dob_close_css_select'])
+        close_dob_filter_button.click()
+
+        sleep(1)
+    except:
+        pass
+
+# TODO: randomly will open "quick view" menu, causes crashes
 # TODO: turn sleeps into sophisticated waits
 # https://selenium-python.readthedocs.io/waits.html#explicit-waits
 if __name__ == '__main__':
@@ -200,9 +274,10 @@ if __name__ == '__main__':
     chrome_options = webdriver.ChromeOptions()
     prefs = {
         'download.default_directory': config['pf_pdf_download_location'],
-        'profile.default_content_settings.popups': True,
+        'profile.default_content_settings.popups': False,
     }
     chrome_options.add_experimental_option('prefs', prefs)
+    chrome_options.add_argument('log-level=3')
     driver = webdriver.Chrome(options=chrome_options)
     driver.maximize_window()
     driver.get(config['pf_url'])
@@ -211,15 +286,25 @@ if __name__ == '__main__':
     go_to_charts(driver)
     sleep(1) # give time for webpage to load
 
-    for row_index in range(2, patient_data.max_row + 1):
+    for row_index in range(198, patient_data.max_row + 1):
+        close_patient_charts_tab(driver)
+
         patient_dob: str = get_date_of_birth(patient_data, row_index)
         enter_date_of_birth(driver, patient_dob)
         
         patient_name: str = get_patient_full_name(patient_data, row_index)
-        navigate_to_patient_info(driver, patient_name)
+        navigation_successful: bool = navigate_to_patient_info(driver, patient_name)
+        if not navigation_successful:
+            print(f'Unable to find patient: {patient_name} (excel row: {row_index})')
+            continue
 
         navigate_to_profile_tab(driver)
-        address_1, address_2, city, state, zip_code = get_patient_address(driver)
+        address_results = get_patient_address(driver)
+        if address_results == False:
+            print(f'Unable to find address for: {patient_name} (excel row: {row_index})')
+            continue
+
+        address_1, address_2, city, state, zip_code = address_results
         patient_data.cell(row_index, config['column_name_mapping']["ADDRESS LINE 1"], address_1)
         patient_data.cell(row_index, config['column_name_mapping']["ADDRESS LINE 2"], address_2)
         patient_data.cell(row_index, config['column_name_mapping']["CITY"], city)
@@ -232,19 +317,19 @@ if __name__ == '__main__':
         navigate_to_documents_tab(driver)
         operation_date: datetime.datetime = patient_data.cell(row_index, config['column_name_mapping']['DATE']).value
         operation_type: str = patient_data.cell(row_index, config['column_name_mapping']['PROCEDURE']).value
-        navigate_to_desired_document(driver, operation_type, operation_date)
-        download_document(driver, patient_name, operation_date, operation_type)
+        navigation_successful = navigate_to_desired_document(driver, operation_type, operation_date)
+        if not navigation_successful:
+            print(f'Unable to find {operation_type} document for {patient_name} on {operation_date.strftime("%m/%d")} (excel row: {row_index})')
+            continue
 
-    #   convert pdf to image
-    #   character recognition on the image
-    #   split resulting string on new line (\n)
-    #   for each line in the pdf
-    #       search for "ID:", "Procedure Code:", "Diagnosis code"
-    #       parse found lines for desired information
-    #   
-    #   add id, procedure code, and diagnosis code to the excel sheet
-        pass
+        pdf_file_path: str = download_document(driver, patient_name, operation_date, operation_type)
+        pdf_text: str = get_text_from_pdf(pdf_file_path)
+        id, procedure_code, diagnosis_code = parse_pdf_text(pdf_text)
 
-    # save the dataframe
+        patient_data.cell(row_index, config['column_name_mapping']["PROCEDURE ID"], id)
+        patient_data.cell(row_index, config['column_name_mapping']["PROCEDURE CODE"], procedure_code)
+        patient_data.cell(row_index, config['column_name_mapping']["DIAGNOSIS"], diagnosis_code)
+
+        workbook.save(f'data/{config["cleaned_workbook_name"]}')
 
     driver.close()
